+++ src/index.ts
@@ -1,4 +1,51 @@
-import { SBVRBoolean }             from './types/Boolean';
+declare global {
+	interface SBVRType<I,O> {
+		types: ConcreteTypes;
+	
+		validate(value: any, required: boolean, cb: Callback<I>): void;
+	
+		fetchProcessing?(data: I, cb: Callback<O>): void;
+	
+		nativeProperties?: NativeProperties;
+	
+		nativeFactTypes?: NativeFactTypes;
+	
+		compare?(data: O, encrypted: I): PromiseLike<boolean>;
+	}
+
+	type Callback<T> = (err?: any, data?: T) => void;
+	type Delayed<T> = (result: T, done: Function) => void;
+
+	type MbDelayed<T> = Delayed<T> | T | Error;
+
+	interface ConcreteTypes {
+		postgres: DBTypeBuilder;
+		mysql: DBTypeBuilder;
+		websql: DBTypeBuilder;
+		odata: ODataTypeBuilder;
+	}
+	
+	type DBTypeBuilder =
+	string
+	| ((necessity: string, index: string) => string);
+	
+	interface ODataTypeBuilder {
+		name: string;
+		complexType?: string;
+	}
+
+	type InternalDate = Date | string | number | null;
+	type NullableDate = Date | null;
+
+	interface RGBA {
+		r: number;
+		g: number;
+		b: number;
+		a: number;
+	}
+}
+
+import { SBVRBoolean }         from './types/Boolean';
 import { BigInteger }          from './types/Big Integer';
 import { CaseInsensitiveText } from './types/Case Insensitive Text';
 import { Color }               from './types/Color';
@@ -18,22 +65,24 @@ import { ShortText }           from './types/Short Text';
 import { Text }                from './types/Text';
 import { Time }                from './types/Time';
 
-exports['Boolean'] = SBVRBoolean;
-exports['Big Integer'] = BigInteger;
-exports['Case Insensitive Text'] = CaseInsensitiveText;
-exports['Color'] = Color;
-exports['ConceptType'] = ConceptType;
-exports['Date'] = SBVRDate;
-exports['Date Time'] = DateTime;
-exports['File'] = File;
-exports['ForeignKey'] = ForeignKey;
-exports['Hashed'] = Hashed;
-exports['Integer'] = Integer;
-exports['Interval'] = Interval;
-exports['JSON'] = SBVRJSON;
-exports['Real'] = Real;
-exports['Serial'] = Serial;
-exports['SHA'] = SHA;
-exports['Short Text'] = ShortText;
-exports['Text'] = Text;
-exports['Time'] = Time;
+export default {
+	'Boolean': SBVRBoolean,
+	'Big Integer': BigInteger,
+	'Case Insensitive Text': CaseInsensitiveText,
+	'Color': Color,
+	'ConceptType': ConceptType,
+	'Date': SBVRDate,
+	'Date Time': DateTime,
+	'File': File,
+	'ForeignKey': ForeignKey,
+	'Hashed': Hashed,
+	'Integer': Integer,
+	'Interval': Interval,
+	'JSON': SBVRJSON,
+	'Real': Real,
+	'Serial': Serial,
+	'SHA': SHA,
+	'Short Text': ShortText,
+	'Text': Text,
+	'Time': Time,
+};
