+++ src/types/JSON.ts
