+++ src/types/Time.ts
@@ -12,7 +12,7 @@ export const Time: SBVRType<string | null, NullableDate> = {
 
 	fetchProcessing: (data, callback) => {
 		let date: NullableDate;
-		if (data !== null) {
+		if (data != null) {
 			// We append the date of the epoch so that we can parse this as a valid date.
 			date = new Date('Thu, 01 Jan 1970 ' + data);
 		} else {
