+++ test/Hashed.ts
