+++ test/Integer.ts
